What: Merge latest changes from 2012 Where: •	Device supported extensions •	Failure of clGetProcAddress(NULL) •	Memory leaks in execution_module.cpp •	Invalid pointer dereferencing •	Changes in framework/api_test.cpp to validate the changes
Test: Pre-commit.

+++ opencl-intel/framework/Execution/sync_graphics_api_objects.cpp
@@ -28,7 +28,7 @@ namespace Intel { namespace OpenCL { namespace Framework
      *      ocl_entry_points * pOclEntryPoints, GraphicsApiMemoryObject** pMemObjects)
      */
 
-    SyncGraphicsApiObjects::SyncGraphicsApiObjects(cl_command_type cmdType, unsigned int uiMemObjNum,
+    SyncGraphicsApiObjects::SyncGraphicsApiObjects(cl_command_type cmdType, size_t uiMemObjNum,
         IOclCommandQueueBase* cmdQueue, ocl_entry_points * pOclEntryPoints,
         GraphicsApiMemoryObject** pMemObjects,bool bIsAcquireCmd) :
       RuntimeCommand(cmdQueue, pOclEntryPoints), m_bIsAcquireCmd(bIsAcquireCmd), m_cmdType(cmdType), m_uiMemObjNum(uiMemObjNum)
